Refactored reports AB#9940
